consumer/requirements.txt
kafka-python==2.0.2
mariadb==1.1.8
packaging==23.2